+++ web/src/lib/storageKeys.js
@@ -1,0 +1,9 @@
+const STORAGE_PREFIX = '@2CPD'
+
+const buildStorageKey = (suffix) => `${STORAGE_PREFIX}/${suffix}`
+
+export const STORAGE_KEYS = {
+  authEmail: buildStorageKey('auth_email'),
+}
+
+export { STORAGE_PREFIX, buildStorageKey }